Updated version of video analysis, see description for detail
Hey Christian,

So it sort of works now.. We however added some stuff to the persons class (line 7, line 41 and everything after line 71). Otherwise we coudnt get it to work :) There is enough room for improvement, but i think it works like this. Not sure if we did it entirely correct though..